PT_ansible/bot_to_git/requirements.txt
aiogram==2.17
python-dotenv
psycopg2-binary
paramiko